changed occ to Pair of uint16_t + uint32_t

+++ find2_index_approx_unidirectional.h
@@ -16,10 +16,10 @@ inline void get_bitvector_interval_inside(Iter<Index<TText, FMIndex<void, TConfi
                                           vector<pair<TVector, TVSupport>> & bitvectors,    
                                           OptimalSearch<nbrBlocks> const & s,
                                           uint8_t const blockIndex,
-                                          Pair<uint8_t, Pair<uint32_t, uint32_t>> & brangeOutput,
+                                          Pair<uint8_t, Pair<uint16_t, uint32_t>> & brangeOutput,
                                           TDir const & ) 
 {
-    Pair<uint32_t, uint32_t> dirrange = range(iter);
+    Pair<uint16_t, uint32_t> dirrange = range(iter);
     uint8_t needed_bitvector;
     uint8_t size = s.pi.size();
     uint8_t bitvsize = bitvectors.size();
@@ -31,7 +31,7 @@ inline void get_bitvector_interval_inside(Iter<Index<TText, FMIndex<void, TConfi
     uint32_t nseq = countSequences(*iter.index);
     dirrange.i1 = dirrange.i1 - nseq;
     dirrange.i2 = dirrange.i2 - nseq;
-//     Pair<uint8_t, Pair<uint32_t, uint32_t>> brangeOutput(needed_bitvector, dirrange);
+//     Pair<uint8_t, Pair<uint16_t, uint32_t>> brangeOutput(needed_bitvector, dirrange);
     brangeOutput.i1 = needed_bitvector;
     brangeOutput.i2 = dirrange;
 }
@@ -91,7 +91,7 @@ inline void uniDirectSearch(TDelegateD & delegateDirect,
                   uint8_t const errors,
                   OptimalSearch<nbrBlocks> const & s,
                   uint8_t const blockIndex,
-                  Pair<uint8_t, Pair<uint32_t, uint32_t>> const & brange,
+                  Pair<uint8_t, Pair<uint16_t, uint32_t>> const & brange,
                   TDir const & /**/)
 {
     auto const & genome = indexText(*iter.index);
@@ -232,7 +232,7 @@ inline void _optimalSearchSchemeExact(TDelegate & delegate,
 
 /* 
 inline ReturnCode checkInterval(vector<pair<sdsl::bit_vector, sdsl::rank_support_v<>>> & bitvectors,
-                          Pair<uint8_t, Pair<uint32_t, uint32_t>> & brange,
+                          Pair<uint8_t, Pair<uint16_t, uint32_t>> & brange,
                           uint8_t const blockSize,
                           bool const done,
                           bool const true,
